support Z data member for points/edges with Clipper 6.0.0
Clipper 6.0.0 alpha introduces a compile time option to support
a Z value for each point. Z is arbitrary data that Clipper passes
through untouched. At intersections it's set to zero, but we can
provide a callback to set it to something else, based on neighboring
Z points. This commit adapts Math::Clipper to support Z, and
introduces an initial edge tracking feature that identifies the
source polygon that contributed edges to portions of result polygons.